chore: move /chat related metadata to chat root layout
